Fixed cart module issue and now Cart module fully done, it is perfectly working when user is loged in
