+++ client/src/api/axiosInstance.js
@@ -1,0 +1,11 @@
+import axios from 'axios'
+
+const axiosInstance = axios.create({
+  baseURL: import.meta.env.VITE_API_URL,  // picks base url value based on env.local or env.production
+  headers: {
+    'Content-Type': 'application/json',
+  },
+  withCredentials: true,
+})
+
+export default axiosInstance
